started to refactor and norm the code. Built-ins, helpers done. created two new files to take the new functions
